requirements.txt
google-cloud==0.34.0
google-api-core==1.16.0
google-auth==1.10.0
googleapis-common-protos==1.6.0
grpc-google-iam-v1==0.12.3
google-cloud-pubsub==1.1.0
google-cloud-kms==1.0.0
google-cloud-storage==1.16.1
google-cloud-error-reporting==0.30.1
protobuf==3.11.2
#html-sanitizer==1.7.2
#html2markdown==0.1.7
flask==1.0.2
psycopg2==2.8.4
#pyjwt[crypto]==1.7.1
pytz==2019.3
PyYAML==5.1
requests==2.22.0
#SQLAlchemy==1.3.4

